Restructure Routine edit page into tabs with per-tab autosave
The Routine edit page was one very tall single form that was hard to
scroll. Split it into Details / Day Entries / Status Criteria tabs that
line up with the View page, mirroring the Domain edit page:

- Details and Status Criteria tabs each own a form and autosave via a
  partial routine upsert (omitted fields are preserved by the backend
  merge), reporting unsaved state up to a single UnsavedChangesDialog.
- Day Entries reuses the self-saving DailyCompletionsManager.
- Bottom action bar uses EditPageFooter with Done / Duplicate / Delete.
- New routines use a minimal create form (name + type) that carries the
  existing search-param prefills, then redirects into the tabbed editor.
- Quick Fill dropdowns use modal={false} to avoid Radix scroll-lock leaks.

The active tab persists in the URL via ?tab=details|entries|criteria.
